# docs/swagger.yaml
definitions:
...
  dto.CharacterCollectionResponse:
    properties:
      characters:
        items:
          $ref: '#/definitions/dto.CharacterResponse'
        type: array
      total:
        type: integer
      unlocked:
        type: integer
    type: object
  dto.CharacterResponse:
    properties:
      achievements:
        items:
          type: string
        type: array
      birth_year:
        type: integer
      death_year:
        type: integer
      description:
        type: string
      dynasty:
        type: string
      era:
        type: string
      famous_quote:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_unlocked:
        type: boolean
      lesson_count:
        type: integer
      name:
        type: string
      rarity:
        type: string
    type: object
  dto.CheckLessonStatusRequest:
    properties:
      lesson_id:
        type: string
    required:
    - lesson_id
    type: object
  dto.CheckLessonStatusResponse:
    properties:
      can_still_pass:
        type: boolean
      earned_points:
        type: integer
      min_score:
        type: integer
      passed:
        type: boolean
      points_needed:
        type: integer
      questions_answered:
        type: integer
      questions_total:
        type: integer
      remaining_points:
        type: integer
      score:
        type: integer
      total_points:
        type: integer
    type: object
...
  dto.CompleteLessonRequest:
    properties:
      lesson_id:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
      time_spent:
        minimum: 1
        type: integer
    required:
    - lesson_id
    - score
    - time_spent
    type: object
...
  dto.CreateSessionRequest:
    properties:
      device_id:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - device_id
    type: object
...
  dto.LessonAccessResponse:
    properties:
      can_access:
        type: boolean
      hearts_needed:
        type: integer
      reason:
        type: string
    type: object
...
  dto.LessonResponse:
    properties:
      animation_status:
        type: string
      animation_url:
        type: string
      audio_status:
        type: string
      audio_url:
        type: string
      can_skip_after:
        description: Content Settings
        type: integer
      character:
        $ref: '#/definitions/dto.CharacterResponse'
      character_id:
        type: string
      has_subtitles:
        type: boolean
      id:
        type: string
      min_score:
        type: integer
      order:
        type: integer
      questions:
        items:
          $ref: '#/definitions/dto.QuestionResponse'
        type: array
      script:
        type: string
      script_status:
        description: Production Workflow
        type: string
      story:
        type: string
      subtitle_url:
        description: Supporting Media
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      xp_reward:
        type: integer
    type: object
...
  dto.QuestionResponse:
    properties:
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      options:
        items:
          type: string
        type: array
      points:
        type: integer
      question:
        type: string
      type:
        type: string
    type: object
...
  dto.SearchResponse:
    properties:
      characters:
        items:
          $ref: '#/definitions/dto.CharacterResponse'
        type: array
      total:
        type: integer
    type: object
...
  dto.SubmitQuestionAnswerRequest:
    properties:
      answer: {}
      lesson_id:
        type: string
      question_id:
        type: string
    required:
    - answer
    - lesson_id
    - question_id
    type: object
  dto.SubmitQuestionAnswerResponse:
    properties:
      can_still_pass:
        type: boolean
      correct:
        type: boolean
      current_score:
        type: integer
      earned_points:
        type: integer
      passed:
        type: boolean
      points:
        type: integer
      points_needed:
        type: integer
      total_points:
        type: integer
    type: object
...
  dto.ValidateLessonRequest:
    properties:
      lesson_id:
        type: string
      user_answers:
        additionalProperties: true
        type: object
    required:
    - lesson_id
    - user_answers
    type: object
  dto.ValidateLessonResponse:
    properties:
      min_score:
        type: integer
      passed:
        type: boolean
      score:
        type: integer
      total_points:
        type: integer
    type: object
...
  shared.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/content/characters:
    get:
      consumes:
      - application/json
      description: Get list of historical characters with filtering options
      parameters:
      - description: Filter by dynasty
        in: query
        name: dynasty
        type: string
      - description: Filter by rarity
        in: query
        name: rarity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterCollectionResponse'
              type: object
      summary: Get Characters
      tags:
      - content
  /api/v1/content/characters/{characterId}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific character
      parameters:
      - description: Character ID
        in: path
        name: characterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterResponse'
              type: object
      summary: Get Character
      tags:
      - content
  /api/v1/content/characters/{characterId}/lessons:
    get:
      consumes:
      - application/json
      description: Get all lessons for a specific character
      parameters:
      - description: Character ID
        in: path
        name: characterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LessonResponse'
                  type: array
              type: object
      summary: Get Character Lessons
      tags:
      - content
...
  /api/v1/content/lessons/{lessonId}:
    get:
      consumes:
      - application/json
      description: Get detailed lesson content including questions
      parameters:
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LessonResponse'
              type: object
      summary: Get Lesson
      tags:
      - content
  /api/v1/content/lessons/questions/answer:
    post:
      consumes:
      - application/json
      description: Submit answer for individual question in a lesson
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Question answer request
        in: body
        name: submitRequest
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitQuestionAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubmitQuestionAnswerResponse'
              type: object
      security:
      - Bearer: []
      summary: Submit Question Answer
      tags:
      - content
  /api/v1/content/lessons/status:
    post:
      consumes:
      - application/json
      description: Check current lesson completion status and score
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson status request
        in: body
        name: statusRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CheckLessonStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CheckLessonStatusResponse'
              type: object
      security:
      - Bearer: []
      summary: Check Lesson Status
      tags:
      - content
  /api/v1/content/lessons/validate:
    post:
      consumes:
      - application/json
      description: Validate user answers for a lesson and return score
      parameters:
      - description: Validation request
        in: body
        name: validateRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ValidateLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ValidateLessonResponse'
              type: object
      summary: Validate Lesson Answers
      tags:
      - content
  /api/v1/content/search:
    get:
      consumes:
      - application/json
      description: Search characters and content with various filters
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: Filter by dynasty
        in: query
        name: dynasty
        type: string
      - description: Filter by rarity
        in: query
        name: rarity
        type: string
      - description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SearchResponse'
              type: object
      summary: Search Content
      tags:
      - content
...
  /api/v1/guest/session:
    post:
      consumes:
      - application/json
      description: This endpoint creates a new guest session or retrieves an existing
        one based on device ID
      parameters:
      - description: Create session request
        in: body
        name: createSessionRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create or Get Guest Session
      tags:
      - guest
  /api/v1/guest/session/{sessionId}/hearts/add:
    post:
      consumes:
      - application/json
      description: This endpoint adds hearts to a guest session when an ad is watched
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Add Hearts from Ad
      tags:
      - guest
  /api/v1/guest/session/{sessionId}/hearts/lose:
    post:
      consumes:
      - application/json
      description: This endpoint deducts a heart from a guest session when a lesson
        is failed
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Lose Heart
      tags:
      - guest
  /api/v1/guest/session/{sessionId}/lesson/{lessonId}/access:
    get:
      consumes:
      - application/json
      description: This endpoint checks if a guest session can access a specific lesson
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LessonAccessResponse'
              type: object
      summary: Check Lesson Access
      tags:
      - guest
  /api/v1/guest/session/{sessionId}/lesson/complete:
    post:
      consumes:
      - application/json
      description: This endpoint marks a lesson as completed for a guest session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Complete lesson request
        in: body
        name: completeLessonRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CompleteLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Complete Lesson
      tags:
      - guest
  /api/v1/guest/session/{sessionId}/progress:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves the progress of a guest session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get Guest Progress
      tags:
      - guest
...
swagger: "2.0"